Fix handling of offsets in load and store
This required several commits. This is really just the latest one. What we tried to achieve is to make sure when using an instruction such as LOAD Rd, Ra, k that the k is calculated as an offset from address 0000 rather than PC, unless k is a label with a dot. In this case the offset is from the nearest non dot label