Changed SHFT to LSH
Motivation: Changed my mind. All assembly code will have a left and right shift operation. Might as well learn that right away. Also it will not be obvious which way the shift operation works when it is just called SHFT. Make it explicit.

This also mirrors well what I do with ADDI and SUBI. We can  just make a RSH pseudo instruction that flips the sign.