Renamed MOVE instruction to LODI
Motivation: In other assembly languages move tends to be a much more generic instruction. In our case it is just Load Immediate. It is thus potentially confusing to call it MOVE.

I discovered this when working on the evil twin of Calcutron-33 called Complexatron-13 shortened to Plexatron-13 which implement a CISC instruction set where the MOVE instruction (or MOV) is quite central.